added base contrast encoder
